requirements.txt
psycopg2-binary==2.8.3 
django-heroku==0.3.1
gunicorn==19.9.0
python-dateutil==2.8.0
pathlib==1.0.1
dj-database-url==0.5.0
python-dateutil==2.8.0
django-import-export==1.2.0
djangorestframework==3.10.2
django-rest-auth==0.9.5
xhtml2pdf==0.2.3
